Allow Advanced Logic to reference specific Nodes
This allows us to ensure that a given NodeTree is only ever exported once.
A node reference acts as a way to attach a plMultiModifier to a
plSceneObject without using those terms. Things will work just fine if the
node reference isn't used, so as long as the entire tree is only used
once.

Future work: ensuring the whole logic tree is only ever exported once.