Fix parsing of Python scripts which use variable Booleans.
Some of Cyan's old scripts predate Python supporting Booleans,
which leaves us with variables where we expected constant values.
This change intercepts these cases and replaces them, avoiding an
exception in Korman when it tries to shove one of these strings
into a boolean attribute.

Also included is a new visitor for proper name constants, support
for which was added in Python 3.4.  This will allow future scripts
to use them and be recognized correctly.