Fix #169.
This fixes the issue in which python nodes reused in multiple pages
won't get properly exported. The exporter doesn't store which page a
node was exported to, and the "optimization" to prevent dupe python
parameters was over-optimizing in the case of multiple pages. Trivial
attempts to track the pages we exported to failed, so we'll fall back to
checking the PFM's filename. If it's None, then obviously the PFM has
not been exported...