Change installer from WiX to NSIS.
I don't really remember why we initially used WiX. Possibly to silently
update any existing Blender installs? Anyway, the problem is that the
VCRuntime MergeModules are deprecated, and GHA has pretty much removed
them. So, we need to use NSIS to be able to trivially run the
vc.redist.exe installer. It runs an MSI, and WiX produces MSIs, so to
keep using WiX, we'd need to write an installer-coordinator. Forget
about it. Just use NSIS.