Improve vertex color exporting significantly.
This moves the piece-by-piece assembly of vertex colors from
`_export_geometry()` into a one-stop-shop for getting the near final
vertex colors as Blender knows them. Included in this is separating out
the adjustment channels for wavesets - which are stuffed inside of
vertex colors. It is now an error for a waveset to have a "col",
"color", or "colour" vertex color layer. This is to prevent confusion.
Wavesets now accept alpha (red), specularity (green), fresnel (blue),
and edgelength (alpha) vertex color layers. The color values in these
layers is averaged and output to the respective channels. Further, a
default value for edgelength is now computed similar (but not exactly
like) PlasmaMax's `SetWaterColor()` function. Artist input to the
edgelength vertex color layer will modulate Korman's calculation.