Don't overwrite trimesh vertices with hull vertices.
If we're given a flat convex hull, we should *only* export the triangle
mesh. The convex hull setting is no longer relevant, and we should
discontinue that code path. Without this return statement, the convex
hull's vertices overwrite the triangle mesh's vertices, but the triangle
mesh indices remain. This can unlease an even worse variety of Satan
that will crash the gane.