Migrate duplicated bounds properties.
This migrates duplicated bounds properties (i.e. those *NOT* found on
the collision modifier) to be stored on the collision modifier. This can
potentially be a lossy process if there are multiple conflicting bounds
property values for a single object. This may result in some Ages having
"new" bugs appear, but, really, they were only working by sheer dumb
luck before. It's better this way.