Fix GUI Camera FOV Default.
For some reason, the operator was using a default of 90 degrees instead of 45 degrees like the rest of the code. This is bad because the result in Blender does not match the in game result with a 90 degree FOV. I suspect this is either a Plasma issue or we need to change the way we output the FOV values. Later.