Add `pfGUIDraggableMod` support to the Draggable Modifier.
I'm not 100% sure what the difference is between DragBar and Draggable
is in terms of when and how they are used. Draggables have some extra
reporting features and can be used by buttons. Unfortunately, there is
only *one* Draggable in all of Uru... the pod map in the Ae'gura Museum,
and it's not a draggable button.

Even though I don't quite understand how it all fits together, I think
it's fine to simply add the support - this Game GUI stuff is still all
explicitly "experimental," so here be dragons, yo.